Adds fixes

+++ scripts/genome_wide/channel_maker_train_germline.py
@@ -459,11 +459,11 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                         # for clipped_arrangement in ['c2c', 'nc2c', 'c2nc', 'nc2nc']:
                         for clipped_arrangement in ['left', 'right', 'all']:
                             clipped_read_distance_array[sample][direction][clipped_arrangement] = np.zeros(win_len,
-                                                                                                           dtype=int)
+                                                                                                           dtype=np.uint32)
                             clipped_read_distance_num[sample][direction][clipped_arrangement] = np.zeros(win_len,
-                                                                                                         dtype=int)
+                                                                                                         dtype=np.uint32)
                             clipped_read_distance_median[sample][direction][clipped_arrangement] = np.zeros(win_len,
-                                                                                                            dtype=int)
+                                                                                                            dtype=np.uint32)
                             for pos in range(start_win, end_win):
                                 if pos in clipped_read_distance[sample][chrName][direction][clipped_arrangement].keys():
                                     clipped_read_distance_array[sample][direction][clipped_arrangement][
@@ -484,7 +484,7 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     # clipped reads
                     clipped_reads_array[sample] = dict()
                     for split_direction in ['left', 'right', 'D_left', 'D_right', 'I']:
-                        clipped_reads_array[sample][split_direction] = np.zeros(win_len, dtype=int)
+                        clipped_reads_array[sample][split_direction] = np.zeros(win_len, dtype=np.uint32)
                         for pos in range(start_win, end_win):
                             if pos in clipped_reads[sample][chrName][split_direction].keys():
                                 clipped_reads_array[sample][split_direction][pos - start_win] = \
@@ -493,7 +493,7 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     # clipped reads inversions
                     clipped_reads_inversion_array[sample] = dict()
                     for mate_position in ['before', 'after']:
-                        clipped_reads_inversion_array[sample][mate_position] = np.zeros(win_len, dtype=int)
+                        clipped_reads_inversion_array[sample][mate_position] = np.zeros(win_len, dtype=np.uint32)
                         for pos in range(start_win, end_win):
                             if pos in clipped_reads_inversion[sample][chrName][mate_position].keys():
                                 clipped_reads_inversion_array[sample][mate_position][pos - start_win] = \
@@ -502,7 +502,7 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     # clipped reads duplication
                     clipped_reads_duplication_array[sample] = dict()
                     for mate_position in ['before', 'after']:
-                        clipped_reads_duplication_array[sample][mate_position] = np.zeros(win_len, dtype=int)
+                        clipped_reads_duplication_array[sample][mate_position] = np.zeros(win_len, dtype=np.uint32)
                         for pos in range(start_win, end_win):
                             if pos in clipped_reads_duplication[sample][chrName][mate_position].keys():
                                 clipped_reads_duplication_array[sample][mate_position][pos - start_win] = \
@@ -511,7 +511,7 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     # clipped reads translocation
                     clipped_reads_translocation_array[sample] = dict()
                     for orientation in ['opposite', 'same']:
-                        clipped_reads_translocation_array[sample][orientation] = np.zeros(win_len, dtype=int)
+                        clipped_reads_translocation_array[sample][orientation] = np.zeros(win_len, dtype=np.uint32)
                         for pos in range(start_win, end_win):
                             if pos in clipped_reads_translocation[sample][chrName][orientation].keys():
                                 clipped_reads_translocation_array[sample][orientation][pos - start_win] = \
@@ -529,9 +529,9 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     split_read_distance_median[sample] = dict()
 
                     for split_direction in ['left', 'right']:
-                        split_read_distance_array[sample][split_direction] = np.zeros(win_len, dtype=int)
-                        split_read_distance_num[sample][split_direction] = np.zeros(win_len, dtype=int)
-                        split_read_distance_median[sample][split_direction] = np.zeros(win_len, dtype=int)
+                        split_read_distance_array[sample][split_direction] = np.zeros(win_len, dtype=np.uint32)
+                        split_read_distance_num[sample][split_direction] = np.zeros(win_len, dtype=np.uint32)
+                        split_read_distance_median[sample][split_direction] = np.zeros(win_len, dtype=np.uint32)
 
                         if pos in split_read_distance[sample][chrName][split_direction].keys():
                             split_read_distance_array[sample][split_direction][pos - start_win] = \
@@ -544,7 +544,7 @@ def channel_maker(ibam, chrList, sampleName, trainingMode, SVmode, outFile):
                     # split reads
                     split_reads_array[sample] = dict()
                     for split_direction in ['left', 'right']:
-                        split_reads_array[sample][split_direction] = np.zeros(win_len, dtype=int)
+                        split_reads_array[sample][split_direction] = np.zeros(win_len, dtype=np.uint32)
                         for pos in range(start_win, end_win):
                             if pos in split_reads[sample][chrName][split_direction].keys():
                                 split_reads_array[sample][split_direction][pos - start_win] = \
